+++ azure/durable_functions/models/OrchestratorState.py
@@ -4,8 +4,8 @@ from typing import List, Any, Dict, Optional
 from azure.durable_functions.models.ReplaySchema import ReplaySchema
 
 from .utils.json_utils import add_attrib
+from .utils.df_serialization import _get_serialize_default
 from azure.durable_functions.models.actions.Action import Action
-from azure.functions._durable_functions import _serialize_custom_object
 
 
 class OrchestratorState:
@@ -114,4 +114,4 @@ class OrchestratorState:
             The instance of the object in json string format
         """
         json_dict = self.to_json()
-        return json.dumps(json_dict, default=_serialize_custom_object)
+        return json.dumps(json_dict, default=_get_serialize_default())